+++ nx-datasets/lib/dataset_utils.ml
@@ -51,21 +51,19 @@ let mkdir_p path perm =
          initial_prefix components);
     ()
 
-let get_cache_base_dir () =
-  match Sys.getenv_opt "NX_DATASETS_CACHE" with
+let get_cache_base_dir ?(getenv = Sys.getenv_opt) () =
+  match getenv "RAVEN_CACHE_ROOT" with
   | Some dir when dir <> "" -> dir
   | _ ->
-      let xdg = Xdg.create ~env:Sys.getenv_opt () in
-      Xdg.cache_dir xdg
+      let xdg = Xdg.create ~env:getenv () in
+      Filename.concat (Xdg.cache_dir xdg) "raven"
 
-let get_cache_dir dataset_name =
-  let base = get_cache_base_dir () in
-  let path =
-    Filename.concat base
-      (Filename.concat "raven" (Filename.concat "datasets" dataset_name))
-  in
-  path ^ "/"
-
+let get_cache_dir ?(getenv = Sys.getenv_opt) dataset_name =
+  let base = get_cache_base_dir ~getenv () in
+  let path = List.fold_left Filename.concat base [ "datasets"; dataset_name ] in
+  let sep = Filename.dir_sep.[0] in
+  if path <> "" && path.[String.length path - 1] = sep then path
+  else path ^ Filename.dir_sep
 
 let mkdir_p dir =
   try mkdir_p dir 0o755 with Unix.Unix_error (Unix.EEXIST, _, _) -> ()